requirements_deploy.txt
streamlit>=1.28.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
openai>=0.28.0
requests>=2.31.0